% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bart.R
\name{bayesian_bootstrap}
\alias{bayesian_bootstrap}
\title{Bayesian Bootstrap for Dirichlet weights}
\usage{
bayesian_bootstrap(n)
}
\description{
Bayesian Bootstrap for Dirichlet weights
}
\keyword{internal}
