% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimands.R
\name{estimands}
\alias{estimands}
\alias{mate_c}
\alias{matt_c}
\alias{sate_c}
\alias{satt_c}
\title{Treatment Effect Estimands for Compliers}
\usage{
mate_c(prince_bart_fit)

matt_c(prince_bart_fit)

sate_c(prince_bart_fit, induce_residual_corr = FALSE)

satt_c(prince_bart_fit, induce_residual_corr = FALSE)
}
\arguments{
\item{prince_bart_fit}{A fitted object from \code{prince_BART}.}

\item{induce_residual_corr}{Logical; for sample estimands, whether to induce
residual correlation between potential outcomes (default: FALSE).}
}
\value{
A \code{posterior} summary object with posterior mean, median,
standard deviation, and quantiles.
}
\description{
Internal functions to compute various treatment effect estimands for the complier
stratum from a fitted prince_bart model.
}
\section{Functions}{
\itemize{
\item \code{mate_c()}: Mixed Average Treatment Effect for Compliers (MATE_C)

\item \code{matt_c()}: Mixed Average Treatment Effect on the Treated
for Compliers (MATT_C)

\item \code{sate_c()}: Sample Average Treatment Effect for Compliers (SATE_C)

\item \code{satt_c()}: Sample Average Treatment Effect on the Treated
for Compliers (SATT_C)

}}
\keyword{internal}
