% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_predictions_for_tree}
\alias{get_predictions_for_tree}
\title{Get Predictions for a Single Tree}
\usage{
get_predictions_for_tree(tree, x)
}
\arguments{
\item{tree}{A data.frame representing a single tree structure.}

\item{x}{A matrix of covariates.}
}
\value{
A numeric vector of predictions for each row in x.
}
\description{
Recursive function to traverse a single BART tree and generate predictions.
}
\keyword{internal}
