% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_bart.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot overlap diagnostics}
\usage{
plot_overlap(x, ...)
}
\arguments{
\item{x}{Output from \code{\link[=general_BART_overlap]{general_BART_overlap()}}}

\item{...}{Additional arguments (ignored)}
}
\description{
Plot overlap diagnostics
}
